Consolidate KAS encryption helpers into assertion.go and helpers.go
Move well-known secret-of-life helpers out of assertion_kas.go and
helpers_kas.go into the existing files and rename them with WellKnown
prefixes per review feedback.
